[REFACTOR] - 공통으로 사용하는 svg 변환을 별도 함수로 분리
